# Use chunked hash calculation