--- a/releases.py
+++ b/releases.py
@@ -5,6 +5,32 @@ import json
 import os
 import re
 
+class ChunkedHash():
+    # Calculate hash for chunked data
+    @staticmethod
+    def hash_bytestr_iter(bytesiter, hasher, ashexstr=True):
+        for block in bytesiter:
+            hasher.update(block)
+        return (hasher.hexdigest() if ashexstr else hasher.digest())
+
+    # Read file in blocks/chunks to be memory efficient
+    @staticmethod
+    def file_as_blockiter(afile, blocksize=65536):
+        with afile:
+          block = afile.read(blocksize)
+          while len(block) > 0:
+              yield block
+              block = afile.read(blocksize)
+
+    # Calculate sha256 hash for a file
+    @staticmethod
+    def calculate_sha256(fname):
+        try:
+            return ChunkedHash.hash_bytestr_iter(ChunkedHash.file_as_blockiter(open(fname, "rb")), hashlib.sha256())
+        except:
+            raise
+            return ""
+
 class ReleaseFile():
     def __init__(self, path, url):
         self._path = path
@@ -124,16 +150,12 @@ class ReleaseFile():
                 for i,release in enumerate(sorted(releases)):
                     key = "%s;%s;%s" % (train, build, release)
                     if key not in self.oldhash:
-                      print("Adding: %s" % release)
-                      sha256hash = hashlib.sha256()
-                      with open(os.path.join(path, release), 'rb') as f:
-                          buf = f.read()
-                          sha256hash.update(buf)
-                      file_digest = sha256hash.hexdigest()
-                      file_size = str(os.path.getsize(os.path.join(path, release)))
+                        print("Adding: %s" % release)
+                        file_digest = ChunkedHash().calculate_sha256(os.path.join(path, release))
+                        file_size = str(os.path.getsize(os.path.join(path, release)))
                     else:
-                      file_digest = self.oldhash[key]["sha256"]
-                      file_size = self.oldhash[key]["size"]
+                        file_digest = self.oldhash[key]["sha256"]
+                        file_size = self.oldhash[key]["size"]
 
                     # .tar
                     self.update_json[train]['project'][build]['releases'][i] = {'file': {'name': release}}
@@ -145,16 +167,12 @@ class ReleaseFile():
                     try:
                         key = "%s;%s;%s" % (train, build, image[0])
                         if key not in self.oldhash:
-                          print("Adding: %s" % image[0])
-                          sha256hash = hashlib.sha256()
-                          with open(os.path.join(path, image[0]), 'rb') as f:
-                              buf = f.read()
-                              sha256hash.update(buf)
-                          file_digest = sha256hash.hexdigest()
-                          file_size = str(os.path.getsize(os.path.join(path, image[0])))
+                            print("Adding: %s" % image[0])
+                            file_digest = ChunkedHash().calculate_sha256(os.path.join(path, image[0]))
+                            file_size = str(os.path.getsize(os.path.join(path, image[0])))
                         else:
-                          file_digest = self.oldhash[key]["sha256"]
-                          file_size = self.oldhash[key]["size"]
+                            file_digest = self.oldhash[key]["sha256"]
+                            file_size = self.oldhash[key]["size"]
                         self.update_json[train]['project'][build]['releases'][i]['image'] = {'name': image[0]}
                         self.update_json[train]['project'][build]['releases'][i]['image']['sha256'] = file_digest
                         self.update_json[train]['project'][build]['releases'][i]['image']['size'] = file_size