When calculating the sha256 hash we currently read the entire binary file into memory, which could result in problems when dealing with large image files on memory limited devices.

This change introduces a chunked hash method, which processes the input file in blocks of 64KB. As a bonus, this new method is significantly faster than the read-all-at-once method, presumably as Python doesn't have to repeatedly allocate large chunks of memory to store the file that is to be hashed.

Comparing a directory of 357 Generic tar files (average 225MB each) on the web03 ARM server, the all-at-once and chunked hashing methods calculated identical hashes, with the chunked method being significantly faster:

```
Elapsed Time for 357 files:
Old method: 2620.00 secs, avg = 7.338948 secs
New method: 1641.09 secs, avg = 4.596879 secs
```

I've also included a small change to cleanup some whitespace issues.