@MilhouseVH 2017-10-04
Updated the second commit to use 4-space indentation, and fix a couple of existing places using 2-space indentation.